+++ e2e/vmss.go
@@ -16,6 +16,7 @@ import (
 
 	"github.com/Azure/agentbaker/pkg/agent"
 	"github.com/Azure/agentbaker/pkg/agent/datamodel"
+	aksnodeconfigv1 "github.com/Azure/agentbaker/pkg/proto/aksnodeconfig/v1"
 	"github.com/Azure/agentbakere2e/config"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
@@ -38,7 +39,9 @@ func createVMSS(ctx context.Context, t *testing.T, vmssName string, scenario *Sc
 	ab, err := agent.NewAgentBaker()
 	require.NoError(t, err)
 	if scenario.AKSNodeConfigMutator != nil {
-		nodeBootstrapping, err = ab.GetNodeBootstrappingForScriptless(ctx, scenario.Runtime.AKSNodeConfig, scenario.VHD.Distro, datamodel.AzurePublicCloud)
+		builder := aksnodeconfigv1.NewNBContractBuilder()
+		builder.ApplyConfiguration(scenario.Runtime.AKSNodeConfig)
+		nodeBootstrapping, err = builder.GetNodeBootstrapping()
 		require.NoError(t, err)
 	} else {
 		nodeBootstrapping, err = ab.GetNodeBootstrapping(ctx, scenario.Runtime.NBC)
