+++ pkg/agent/bakerapi.go
@@ -5,8 +5,6 @@ package agent
 
 import (
 	"context"
-	"encoding/base64"
-	"encoding/json"
 	"fmt"
 
 	"github.com/Azure/agentbaker/pkg/agent/datamodel"
@@ -85,50 +83,6 @@ func (agentBaker *agentBakerImpl) GetNodeBootstrapping(ctx context.Context, conf
 	}
 
 	return nodeBootstrapping, nil
-}
-
-// TODO: config supposed to be type *aksnodeconfigv1.Configuration, but can't import it here
-// because of circular dependency.
-func (agentBaker *agentBakerImpl) GetNodeBootstrappingForScriptless(
-	ctx context.Context,
-	config any,
-	distro datamodel.Distro,
-	cloudName string,
-) (*datamodel.NodeBootstrapping, error) {
-	scriptlessCustomData, err := getScriptlessCustomDataContent(config)
-	if err != nil {
-		return nil, err
-	}
-
-	nodeBootstrapping := &datamodel.NodeBootstrapping{
-		CSE:        scriptlessBootstrapStatusCSE,
-		CustomData: scriptlessCustomData,
-	}
-
-	if distro == datamodel.CustomizedWindowsOSImage || distro == datamodel.CustomizedImage || distro == datamodel.CustomizedImageKata {
-		return nodeBootstrapping, nil
-	}
-
-	osImageConfigMap, hasCloud := datamodel.AzureCloudToOSImageMap[cloudName]
-	if !hasCloud {
-		return nil, fmt.Errorf("don't have settings for cloud %s", cloudName)
-	}
-
-	if osImageConfig, hasImage := osImageConfigMap[distro]; hasImage {
-		nodeBootstrapping.OSImageConfig = &osImageConfig
-	}
-
-	return nodeBootstrapping, nil
-}
-
-func getScriptlessCustomDataContent(config any) (string, error) {
-	nbcJSON, err := json.Marshal(config)
-	if err != nil {
-		return "", fmt.Errorf("failed to marshal nbc, error: %w", err)
-	}
-	encodedNBCJson := base64.StdEncoding.EncodeToString(nbcJSON)
-	customDataYAML := fmt.Sprintf(scriptlessCustomDataTemplate, encodedNBCJson)
-	return base64.StdEncoding.EncodeToString([]byte(customDataYAML)), nil
 }
 
 func (agentBaker *agentBakerImpl) GetLatestSigImageConfig(sigConfig datamodel.SIGConfig,
