+++ pkg/proto/aksnodeconfig/v1/const.go
@@ -20,3 +20,14 @@ const (
 	AksCustomCloudName   = "akscustom"
 	AzureStackCloud      = "AzureStackCloud"
 )
+
+const (
+	scriptlessBootstrapStatusCSE = "/opt/azure/containers/aks-node-controller provision-wait"
+	scriptlessCustomDataTemplate = `#cloud-config
+write_files:
+- path: /opt/azure/containers/aks-node-controller-config.json
+  permissions: "0755"
+  owner: root
+  content: !!binary |
+   %s`
+)
